adding PROXY_HEADER_HOST env var
Allow users to set the proxied Host header value.
Otherwise, ssl_proxy assumes this should be set to $host,
which strips off the port number.

An altered `$host` value can break some use cases,
such as when the client signs the headers in the request,
which is the case for e.g., Amazon S3 or
MinIO, when serving on a non-standard port.
In this case, it's important to set the Host header exactly,
or use `$http_host` to preserve the port number.